testbench/pyproject.toml
[project]
name = "testbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cocotb>=2.0.1",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
]
